Enable elevation data use with standalone graph builder.
The existing "elevation" command-line argument attempts to download NED data automatically,
but the implementation for that is broken due to changes in how NED data is served.
This will automatically find GeoTIFF files in the data directory, and use the first one found
to build elevation profiles into the graph.